Optimize Cop#relevant_rubocop_rspec_file?
The result of `rspec_pattern` is always the same (because it is based
on the configuration file), so there is no need to re-calculate it every
time it is called.

I tried memozing on the Cop instance, but at least when running the
specs, we instantiate Cop way too often. Memoizing on the class level
reduces the number of calculations to once per `Cop` subclass we have.
If we allowed using class variables (`@@rspec_pattern`) we could reduce
the number of calculations to one.

This commit reverts 995b4fe5f16cd73feb2625eadd0fad236cfab38e and adds
memoization in a class instance variable.